Adds json tempmlate tests for sig reos in search
This also moves some functions out of the "view" into the json template.
I think this is the right pattern for functions/data mutations that
aren't shared between html and json templates.
